# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления файловым хранилищем",
        "title": "Meemo File Storage API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@meemo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех файлов текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Получить список файлов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.GetAllUserFilesDtoOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/metadata": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает метаданные для нового файла",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Создать метаданные файла",
                "parameters": [
                    {
                        "description": "Метаданные файла",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenter_http_handler_file.SaveFileMetadata"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.SaveFileMetadataDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/rename": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменяет имя файла",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Переименовать файл",
                "parameters": [
                    {
                        "description": "Запрос на переименование",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenter_http_handler_file.RenameFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.RenameFileDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменяет статус файла (например, 0 - активен, 1 - удален)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Изменить статус файла",
                "parameters": [
                    {
                        "description": "Запрос на изменение статуса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenter_http_handler_file.SetStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.SetStatusDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/visibility": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменяет видимость файла (публичный/приватный)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Изменить приватность файла",
                "parameters": [
                    {
                        "description": "Запрос на изменение приватности",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenter_http_handler_file.ChangeVisibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.ChangeVisibilityDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/{id}/content": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает содержимое файла по его ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить содержимое файла",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Содержимое файла",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.SaveFileContentDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Скачивает файл по его имени",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Получить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет файл по его имени",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Удалить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.DeleteFileDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/{name}/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает метаданные файла по его имени",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Получить информацию о файле",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meemo_internal_usecase_file.GetFileInfoDtoOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_presenter_http_handler_file.ChangeVisibilityRequest": {
            "type": "object",
            "properties": {
                "is_public": {
                    "type": "boolean"
                },
                "original_name": {
                    "type": "string"
                }
            }
        },
        "internal_presenter_http_handler_file.RenameFileRequest": {
            "type": "object",
            "properties": {
                "new_name": {
                    "type": "string"
                },
                "old_name": {
                    "type": "string"
                }
            }
        },
        "internal_presenter_http_handler_file.SaveFileMetadata": {
            "type": "object",
            "properties": {
                "is_public": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "size_in_bytes": {
                    "type": "integer"
                }
            }
        },
        "internal_presenter_http_handler_file.SetStatusRequest": {
            "type": "object",
            "properties": {
                "original_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "meemo_internal_usecase_file.ChangeVisibilityDtoOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "original_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "meemo_internal_usecase_file.DeleteFileDtoOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "meemo_internal_usecase_file.FileListItemDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "size_in_bytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "meemo_internal_usecase_file.GetAllUserFilesDtoOut": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meemo_internal_usecase_file.FileListItemDto"
                    }
                }
            }
        },
...
        "meemo_internal_usecase_file.RenameFileDtoOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "new_name": {
                    "type": "string"
                },
                "old_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "meemo_internal_usecase_file.SaveFileContentDtoOut": {
            "type": "object",
            "properties": {
                "loading_result": {
                    "type": "boolean"
                }
            }
        },
        "meemo_internal_usecase_file.SaveFileMetadataDtoOut": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "size_in_bytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "meemo_internal_usecase_file.SetStatusDtoOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "original_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "User ID header",
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        },
        "ApiKeyAuth2": {
            "description": "User Email header",
            "type": "apiKey",
            "name": "X-User-Email",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}